Hogwarts Three SQL

+++ src/main/java/ru/prostostudia/hogwartslegacy/services/FacultyServiceImpl.java
@@ -1,23 +1,24 @@
 package ru.prostostudia.hogwartslegacy.services;
 
 import org.springframework.stereotype.Service;
-import ru.prostostudia.hogwartslegacy.exceptions.FacultyIllegalParameterException;
-import ru.prostostudia.hogwartslegacy.exceptions.FacultyNameSetAlreadyException;
-import ru.prostostudia.hogwartslegacy.exceptions.FacultyNotFoundException;
+import ru.prostostudia.hogwartslegacy.exceptions.*;
 import ru.prostostudia.hogwartslegacy.interfaces.FacultyService;
+import ru.prostostudia.hogwartslegacy.interfaces.StudentService;
 import ru.prostostudia.hogwartslegacy.models.Faculty;
+import ru.prostostudia.hogwartslegacy.models.Student;
 import ru.prostostudia.hogwartslegacy.repository.FacultyRepository;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Service
 public class FacultyServiceImpl implements FacultyService {
 
     private final FacultyRepository facultyRepository;
+    private final StudentService studentService;
 
-    public FacultyServiceImpl(FacultyRepository facultyRepository) {
+    public FacultyServiceImpl(FacultyRepository facultyRepository, StudentService studentService) {
         this.facultyRepository = facultyRepository;
+        this.studentService = studentService;
     }
 
     @Override
@@ -47,6 +48,9 @@ public class FacultyServiceImpl implements FacultyService {
     public void remove(Long id) {
         facultyRepository.findById(id)
                 .orElseThrow(FacultyNotFoundException::new);
+        if(!studentService.filterByFaculty(id).isEmpty()) {
+            throw new FacultyContainStudentException();
+        }
         facultyRepository.deleteById(id);
     }
 
@@ -74,9 +78,16 @@ public class FacultyServiceImpl implements FacultyService {
 
     @Override
     public List<Faculty> filterByColor(String color) {
-        return  facultyRepository.findAll().stream()
-                .collect(Collectors.groupingBy(Faculty::getColor))
-                .get(color);
+        return  facultyRepository.findByColorIgnoreCase(color);
+    }
+
+    @Override
+    public List<Student> getStudents(Long id) {
+        List<Student> students =  studentService.filterByFaculty(id);
+        if (students.isEmpty()) {
+            throw new StudentNotFoundException();
+        }
+        return students;
     }
 
     private void verifyParameter(String name, String color) {
